projects/hcp/packages/rahcp-client/pyproject.toml
[project]
name = "rahcp-client"
version = "0.1.0"
description = "Async Python client for HCP Unified API"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28",
    "pydantic-settings>=2.13",
    "rahcp-tracker",
    "tenacity>=8.1",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-exporter-otlp>=1.29",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.25", "respx>=0.22"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv.sources]
rahcp-tracker = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
